+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 ### Improvements
 * Specialised lattice constructors like {func}`nk.graph.Grid` now accept a `point_group` argument, overriding the default (usually maximal) point groups [#1879](https://github.com/netket/netket/pull/1879).
+* Methods to generate random states are automatically implemented for all {class}`nk.hilbert.HomogeneousHilbert`, constrained or not [#1911](https://github.com/netket/netket/pull/1911).
 
 ### Bug fixes
 * Fix the function {meth}`nk.graph.SpaceGroupBuilder.space_group_irreps` throwing away the imaginary part of point-group characters, which led to incorrect space-group characters in some rare cases [#1876](https://github.com/netket/netket/pull/1876).